ravi-backend/app/services/tts.py
from google.cloud import texttospeech
from app.config import settings

class TTSService:
    def __init__(self):
        self.client = texttospeech.TextToSpeechClient()
        self.voice = texttospeech.VoiceSelectionParams(
            language_code=settings.TTS_LANGUAGE,
            name=settings.TTS_VOICE,
        )
        self.audio_config = texttospeech.AudioConfig(
            audio_encoding=getattr(texttospeech.AudioEncoding, settings.TTS_AUDIO_ENCODING)
        )

    def synthesize(self, text: str) -> bytes:
        synthesis_input = texttospeech.SynthesisInput(text=text)
        response = self.client.synthesize_speech(
            input=synthesis_input,
            voice=self.voice,
            audio_config=self.audio_config,
        )
        return response.audio_content
