ravi-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-texttospeech
Pillow